Changed the way of checking if contact is already present in the group
This change was done due to the fact that we already have the contact in our phone book and to add a contact to the ContactGroup we wouldn't create a new object. The previous way of validating this is more suitable for the PhoneBook class - so that we won't have two exactly the same contacts.